- ITS ALIVE! - Save frames in gro format - LJ potential and verlet integrator

+++ base.h
@@ -1,15 +1,15 @@
 #ifndef BASE_H
 #define BASE_H
 #include <eigen3/Eigen/Dense>
-#include "frame.h"
 #include "atom.h"
+#include "frame.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <iostream>
 
 class Base{
     public:
-        static int box_dim;
+        static double box_dim;
         static int num_of_atoms;
         static int iterations;
 };
